split code_803E46C
